requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
discord.py==1.4.1
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.21.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
mcrcon==0.6.0
multidict==4.7.6
Pillow==7.2.0
praw==7.1.0
prawcore==1.5.0
protobuf==3.13.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dotenv==0.14.0
quantumrandom==1.9.0
requests==2.24.0
rsa==4.6
six==1.15.0
termcolor==1.1.0
update-checker==0.18
uritemplate==3.0.1
urllib3==1.25.10
websocket-client==0.57.0
websockets==8.1
youtube-dl==2020.7.28
asyncio
pytz
timeago==1.0.14
yarl==1.5.1
owotext==1.0.3
pydantic==1.6.1
beautifulsoup4==4.9.1
strawpy==1.1
GitPython==3.1.7
pyfiglet==0.8.post1
tokage==1.4
PyGithub==1.53
aiopypixel==1.8.2
hypixel==0.7.9